divide filters and pagination in sub components
