+++ new_project/src/views/Home.vue
@@ -1,5 +1,6 @@
 <template>
   <div class="home">
+  <h1 class="mainTitle">RICK AND MORTY X JELLYSMACK</h1>
     <ListOfCharacters/>
   </div> 
 </template>
@@ -16,9 +17,22 @@ export default {
 }
 </script> 
 
-<style>
+<style scoped>
 
-
+.mainTitle {
+  margin-top: 5%;
+  margin-bottom: 10%;
+  font-size: 5em;
+  background: linear-gradient(
+    90deg,
+    rgba(131, 58, 180, 1) 0%,
+    rgba(253, 29, 29, 1) 50%,
+    rgba(252, 176, 69, 1) 100%
+  );
+  background-clip: border-box;
+  -webkit-background-clip: text;
+  -webkit-text-fill-color: transparent;
+}
 
 </style>
